Added new toy to testing - INCOMPLETE
Hand-programmed a table of contents with scroll detection and scroll behaviour.
The table of contents is completely scalable, any elements inside, no matter the height, will be detected if they are scrolled to.
Also made the cursor art sitewide and animatable.
These will only work on desktop at the moment.